--- frontend/src/app/core/interceptors/http interceptors/index.ts
@@ -1,9 +1,0 @@
-/* "Barrel" of Http Interceptors */
-import { HTTP_INTERCEPTORS } from '@angular/common/http';
-
-import { ErrorInterceptor } from './error-handler.interseptor';
-
-/** Http interceptor providers in outside-in order */
-export const httpInterceptorProviders = [
-  { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
-];
