+++ frontend/src/app/core/interceptors/index.ts
@@ -1,0 +1,1 @@
+export * from './error-handler.interseptor';
